requirements.txt
scikit-learn==0.23.1
keras==2.2.4
tensorflow==1.15
numpy==1.17.4
pandas
pytest
PyYAML
parse
Flask
jsonpickle
requests
scipy==1.4.1
environs
pathlib2
diffprivlib
psutil
setproctitle
tabulate
lz4
opencv-python
gym
cloudpickle==1.3.0
skorch
torch==1.4.0
h5py==2.10.0
image
aif360
fairlearn
gensim
ray
dm-tree
GPUtil
paramiko
http://github.com/IBM/pycloudmessenger/archive/v0.7.1.tar.gz
jupyter
kubernetes
openshift